Read one file from a Bascet

TODO This can be made faster by, e.g., once and for all reading the location of all objects in the file

Usage

BascetReadFile(
  bascetFile,
  cellID,
  filename,
  as = c("tempfile"),
  bascet_instance = bascet_instance.default,
  verbose = FALSE
)

Arguments

bascetFile

Handle for an opened Bascet file

cellID

ID of the cell (string)

filename

description

as

description

bascet_instance

Configuration for how to run the Bascet Rust API

verbose

description

Value

Name of a temporary file where the read content is stored