############################################### #' @return TODO #' @export LoadFragmentsAsObject <- function(fragpath)

Index fragment file if needed. Only needed if you got the file externally and it has no index

fragpath_index <- paste(fragpath,".tbi",sep="") if(!file.exists(fragpath_index)) print("Indexing fragment file") system(paste("tabix -p vcf ",fragpath))

Figure out names of cells

Create a dummy assay

stupidmat <- matrix(0, nrow = 2, ncol=length(keep_cells)) rownames(stupidmat) <- c("chr1:1-100", "chr1:200-300") colnames(stupidmat) <- keep_cells

chrom_assay <- CreateChromatinAssay( counts = as.sparse(stupidmat), sep = c(":", "-"), fragments = fragpath, min.cells = 0, min.features = 0 )

adata <- CreateSeuratObject( counts = chrom_assay, assay = "aligned" )

adata

############################################### #' @return TODO #' @export LoadFragmentsAsObject <- function(fragpath)

Index fragment file if needed. Only needed if you got the file externally and it has no index

fragpath_index <- paste(fragpath,".tbi",sep="") if(!file.exists(fragpath_index)) print("Indexing fragment file") system(paste("tabix -p vcf ",fragpath))

Figure out names of cells

Create a dummy assay

stupidmat <- matrix(0, nrow = 2, ncol=length(keep_cells)) rownames(stupidmat) <- c("chr1:1-100", "chr1:200-300") colnames(stupidmat) <- keep_cells

chrom_assay <- CreateChromatinAssay( counts = as.sparse(stupidmat), sep = c(":", "-"), fragments = fragpath, min.cells = 0, min.features = 0 )

adata <- CreateSeuratObject( counts = chrom_assay, assay = "aligned" )

adata

Usage

CountGrangeFeatures(adata, grange_gene)

Value

TODO